chore: transfer

+++ src/infrastructure/store/gantt.store.ts
@@ -6,10 +6,23 @@ import { Marco } from '../../domain/entities/marco';
 import { Ganttable } from '../../domain/aggregates/ganttable';
 import { api } from '../boot/axios';
 import { Projeto } from '../../domain/entities/projeto';
+import { Usuario } from '../../domain/entities/usuario';
 
 export const useGanttStore = defineStore('gantt', () => {
     const projeto = ref<Projeto>();
     const projetos = ref<Projeto[]>();
+
+    const responsaveis = computed(() => {
+        const r: Usuario[] = [];
+        tarefas.value?.forEach((tarefa) => {
+            tarefa.usuariosTarefas.forEach((usuario) => {
+                if (!r.includes(usuario)) {
+                    r.push(usuario);
+                }
+            });
+        });
+        return r;
+    });
 
     const marcos = computed(() => {
         const m: Marco[] = [];
@@ -29,14 +42,17 @@ export const useGanttStore = defineStore('gantt', () => {
         return t;
     });
     const ganttables = computed(() => {
+        const marcosOrderByDataInicio = marcos.value?.sort((a, b) => {
+            return (
+                a.tarefas[0].dataInicio.getTime() -
+                b.tarefas[0].dataInicio.getTime()
+            );
+        });
         const g: Ganttable[] = [];
-        marcos.value?.forEach((marco) => {
+        marcosOrderByDataInicio.forEach((marco) => {
             g.push(new Marco(marco));
             const tarefasOrderByDataInicio = marco.tarefas.sort((a, b) => {
-                return (
-                    new Date(a.dataInicio).getTime() -
-                    new Date(b.dataInicio).getTime()
-                );
+                return a.dataInicio.getTime() - b.dataInicio.getTime();
             });
             tarefasOrderByDataInicio.forEach((tarefa) => {
                 g.push(new Tarefa(tarefa));
@@ -46,8 +62,8 @@ export const useGanttStore = defineStore('gantt', () => {
     });
 
     async function buscaDadosGantt(projetoId: string) {
-        const { data } = await api.get<any>('/projeto');
-        projeto.value = new Projeto(data);
+        const { data } = await api.get<Projeto>('/projeto');
+        projeto.value = data;
     }
 
     async function criaTarefa(tarefa: Tarefa) {
@@ -61,18 +77,21 @@ export const useGanttStore = defineStore('gantt', () => {
     }
 
     async function atualizaTarefa(tarefa: Tarefa) {
+        if (!projeto.value) return;
         const { data } = await api.put<Tarefa>(`/tarefas/${tarefa.id}`, tarefa);
-        await buscaDadosGantt(data.projetoId);
+        await buscaDadosGantt(projeto.value.id);
     }
 
     async function atualizaMarco(marco: Marco) {
+        if (!projeto.value) return;
         const { data } = await api.put<Marco>(`/marcos/${marco.id}`, marco);
-        await buscaDadosGantt(data.projetoId);
+        await buscaDadosGantt(projeto.value.id);
     }
 
     async function removeTarefa(tarefa: Tarefa) {
+        if (!projeto.value) return;
         const { data } = await api.delete(`/tarefas/${tarefa.id}`);
-        await buscaDadosGantt(data.projetoId);
+        await buscaDadosGantt(projeto.value.id);
     }
 
     return {
@@ -86,5 +105,6 @@ export const useGanttStore = defineStore('gantt', () => {
         atualizaMarco,
         removeTarefa,
         projeto,
+        responsaveis,
     };
 });
